requirements.txt
# Core dependencies
requests==2.31.0
python-dotenv==1.0.0

# Database
supabase==2.0.3

# OpenAI
openai>=1.0.0

# Email services
sendgrid==6.10.0

# Data visualization
matplotlib==3.8.0

# Optional email service
python-http-client>=3.0.0  # Required by SendGrid

# Schedule
schedule==1.2.0